backend/requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
pydantic==2.8.0
pymongo==4.8.0
Pillow==10.0.0
requests==2.32.3
python-dotenv==1.0.0
google-generativeai==0.5.4
pytz==2023.3
passlib[bcrypt]==1.7.4
python-jose[jwt]==3.3.0
python-multipart==0.0.9